Correction de bug. Version STABLE. Ready to ship.
* APIIndex.cs: Correction d'un bug empechant l'app de détecter les
  fichiers de sujet s'il sont contenus dans un sous-dossier.

* Epitech.Intra.SharedApp.cs: Préparation pour l'utilisation d'un
  prompt de notationde l'app.

* packages.config:
* packages.config:
* Epitech.Intra.iOS.csproj:
* Settings.cs:
* _SettingsStarted.txt:
* Epitech.Intra.SharedApp.csproj: 

* Notifications.cs: Assignation d'une taille fixe aux image de la
  liste des notifications pour prévenir un bug graphique récurrent.

* CreditHeader.cs: Modification esthétique.

* AppDelegate.cs: Petite modification empechant d'effectuer un appel
  système interdit dans certaines conditions.

* EventManager.cs: Correction d'un bug dans le getter du Calendrier
  iOS.
Changement d'appel pour ne plus utiliser une propriété dépréciée.

+++ Epitech.Intra.iOS/DependencyService/EventManager.cs
@@ -13,7 +13,16 @@ namespace Epitech.Intra.iOS
 {
 	public class EventManagerIOS : IEventManagerIOS
 	{
-		public static EKEventStore EventStore;
+		static protected EKEventStore eventStore;
+
+		static public EKEventStore EventStore {
+			get {
+				if (eventStore == null)
+					eventStore = new EKEventStore ();
+				return eventStore; 
+			}
+		}
+
 		const string CalendarIDKey = "CalendarIdForEventSynch";
 
 		public static EKCalendar Calendar {
@@ -22,7 +31,7 @@ namespace Epitech.Intra.iOS
 				if (val == null) {
 					return CreateNewCalendar ();
 				} else {
-					foreach (var item in EventStore.Calendars) {
+					foreach (var item in EventStore.GetCalendars(EKEntityType.Event)) {
 						if (item.CalendarIdentifier == val)
 							return item;
 					}
@@ -60,17 +69,19 @@ namespace Epitech.Intra.iOS
 
 		public async void SynchrosizeCalendar (List<Calendar> events)
 		{
-			try {
+			if ((((App)Xamarin.Forms.Application.Current).HasAllowedEventKit) && (((App)Xamarin.Forms.Application.Current).UserHasActivatedEventSync)) {
 				await Task.Factory.StartNew (new Action (() => {
-					if ((((App)Xamarin.Forms.Application.Current).HasAllowedEventKit) || (((App)Xamarin.Forms.Application.Current).UserHasActivatedEventSync)) {
+					try {
 						SelectUnregisteredEvents (events);
 						RegisterEvents (events);
 						DeleteUnregisteredEvent (events);
+					} catch (Exception ex) {
+						Xamarin.Insights.Report (ex);
+						return;
 					}
 				}));
-			} catch (Exception ex) {
-				Xamarin.Insights.Report (ex);
 			}
+			return;
 		}
 
 		public static void RegisterEvents (List<Calendar> events)
@@ -105,7 +116,7 @@ namespace Epitech.Intra.iOS
 				return;
 			EKEvent[] todelete = Array.FindAll (queryresult, x => {
 				foreach (var item in events) {
-					if (item.EventKitID == x.EventIdentifier)
+					if (item.EventKitID == x.EventIdentifier || item.Past)
 						return false;
 				}
 				return true;
